Allow skipping per-query namespace inference via config
This is a way to open the current behaviour of Loki "meta-data"
queries to other use cases where the list of namespaces cannot be
inferred from queries.

NetObserv will use this flag because its queries are not namespaced.
This allows to effectively use the fine-grained SAR feature with
netobserv.

Related JIRA: NETOBSERV-1324